% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{all_dc_to_numeric}
\alias{all_dc_to_numeric}
\title{Turn all columns with character values with decimal commas into numeric columns with decimal points}
\usage{
all_dc_to_numeric(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A data frame.
}
\description{
Turns "3,78" into 3.78 for all columns containing numeric data with decimal commas stored as text.
}
\details{
Replaces all commas "," with points "." in alla text columns and converts the values to numeric.
}
\examples{
SFM_example |> all_dc_to_numeric()
}
