% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{dc_to_numeric}
\alias{dc_to_numeric}
\title{Turn character values with decimal commas into numeric values with decimal points}
\usage{
dc_to_numeric(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A numeric vector.
}
\description{
Turns "3,78" into 3.78.
}
\details{
Replaces all commas "," with points "." and converts the values to numeric.
}
\examples{
library(dplyr)
SFM_example |> mutate(across(where(is.character), dc_to_numeric))
}
