% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{hydroyear}
\alias{hydroyear}
\title{Compute hydrological year from a date}
\usage{
hydroyear(x)
}
\arguments{
\item{x}{A Date vector, or a character vector containing dates using the YYYY-MM-DD format.}
}
\value{
A numeric vector.
}
\description{
The hydrological year is the same as the calendar year for January-September, and the calendar year +1 for October-December.
}
\examples{
dates <- c("2024-03-15", "2024-08-13", "2024-11-29")
hydroyear(dates)
}
