% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\docType{data}
\name{SFM_example}
\alias{SFM_example}
\title{Example data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 8 columns.
}
\usage{
SFM_example
}
\description{
Example data
}
\keyword{datasets}
